Fix the git file-deletion problem
Since we are working with "bare" repositories, the index either doesn't exist or is out of date most of the time. Therefore, when we add files, we are adding them to a blank index, which of course wipes everything out.

The solution is to do a "git reset" before adding files, getting the index synced up with the current master before proceeding.